fix: 🐞 move AnchorAnnotator into new annotators.py
